# backend/public/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"Board": {
				"properties": {
					"BID": {
						"type": "number",
						"format": "double"
					},
					"BOARDTITLE": {
						"type": "string"
					},
					"BOARDCONTENT": {
						"type": "string"
					},
					"BOARDREGDATE": {
						"type": "string"
					},
					"MEMBERNAME": {
						"type": "string"
					}
				},
				"required": [
					"BOARDTITLE",
					"BOARDCONTENT",
					"MEMBERNAME"
				],
				"type": "object",
				"additionalProperties": false
			},
			"BoardCreationParams": {
				"properties": {
					"BOARDTITLE": {
						"type": "string"
					},
					"BOARDCONTENT": {
						"type": "string"
					},
					"MEMBERNAME": {
						"type": "string"
					}
				},
				"required": [
					"BOARDTITLE",
					"BOARDCONTENT",
					"MEMBERNAME"
				],
				"type": "object",
				"additionalProperties": false
			},
			"BoardUpdateParams": {
				"properties": {
					"BOARDTITLE": {
						"type": "string"
					},
					"BOARDCONTENT": {
						"type": "string"
					},
					"MEMBERNAME": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"IChatRoom": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string",
						"nullable": true
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IChatMessage": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"content": {
						"type": "string"
					},
					"senderId": {
						"type": "string"
					},
					"roomId": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"content",
					"senderId",
					"roomId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IMember": {
				"properties": {
					"MID": {
						"type": "string"
					},
					"MPW": {
						"type": "string"
					},
					"MEMBERNAME": {
						"type": "string",
						"nullable": true
					},
					"MROLE": {
						"type": "string"
					}
				},
				"required": [
					"MID",
					"MEMBERNAME",
					"MROLE"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Error": {
				"properties": {
					"name": {
						"type": "string"
					},
					"message": {
						"type": "string"
					},
					"stack": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"MemberCreationParams": {
				"properties": {
					"MID": {
						"type": "string"
					},
					"MPW": {
						"type": "string"
					},
					"MEMBERNAME": {
						"type": "string",
						"nullable": true
					},
					"MROLE": {
						"type": "string"
					}
				},
				"required": [
					"MID",
					"MPW"
				],
				"type": "object",
				"additionalProperties": false
			},
			"MemberLoginParams": {
				"properties": {
					"MID": {
						"type": "string"
					},
					"MPW": {
						"type": "string"
					}
				},
				"required": [
					"MID",
					"MPW"
				],
				"type": "object",
				"additionalProperties": false
			},
			"MemberUpdateParams": {
				"properties": {
					"MPW": {
						"type": "string"
					},
					"MEMBERNAME": {
						"type": "string"
					},
					"MROLE": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"jwt": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "TSOA",
		"version": "1.0.0",
		"description": "Build swagger-compliant REST APIs using TypeScript and Node",
		"license": {
			"name": "MIT"
		},
		"contact": {
			"name": "unknown"
		}
	},
	"openapi": "3.0.0",
	"paths": {
		"/api/boards": {
			"get": {
				"operationId": "GetBoards",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Board"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Board"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "CreateBoard",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Board"
								}
							}
						}
					}
				},
				"tags": [
					"Board"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/BoardCreationParams"
							}
						}
					}
				}
			}
		},
		"/api/boards/{boardId}": {
			"get": {
				"operationId": "GetBoard",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Board"
								}
							}
						}
					}
				},
				"tags": [
					"Board"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "boardId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateBoard",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Board"
								}
							}
						}
					}
				},
				"tags": [
					"Board"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "boardId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/BoardUpdateParams"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteBoard",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Board"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "boardId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/api/chat/rooms": {
			"get": {
				"operationId": "GetChatRooms",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/IChatRoom"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Chat"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
		"/api/chat/rooms/{roomId}": {
			"get": {
				"operationId": "GetChatRoom",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IChatRoom"
								}
							}
						}
					}
				},
				"tags": [
					"Chat"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "roomId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/chat/rooms/{roomId}/messages": {
			"get": {
				"operationId": "GetChatMessages",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/IChatMessage"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Chat"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "roomId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/members": {
			"get": {
				"operationId": "GetMembers",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/IMember"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Member"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
		"/api/members/{mid}": {
			"get": {
				"operationId": "GetMember",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IMember"
								}
							}
						}
					}
				},
				"tags": [
					"Member"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "mid",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateMember",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IMember"
								}
							}
						}
					}
				},
				"tags": [
					"Member"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "mid",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/MemberUpdateParams"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteMember",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Member"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "mid",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/members/register": {
			"post": {
				"operationId": "RegisterMember",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								},
								"examples": {
									"Example 1": {
										"value": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
									}
								}
							}
						}
					},
					"201": {
						"description": "회원가입 성공",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					},
					"409": {
						"description": "이미 존재하는 사용자",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Error"
								}
							}
						}
					}
				},
				"tags": [
					"Member"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/MemberCreationParams"
							}
						}
					}
				}
			}
		},
		"/api/members/login": {
			"post": {
				"operationId": "LoginMember",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								},
								"examples": {
									"Example 1": {
										"value": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
									}
								}
							}
						}
					},
					"401": {
						"description": "인증 실패",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Error"
								}
							}
						}
					}
				},
				"tags": [
					"Member"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/MemberLoginParams"
							}
						}
					}
				}
			}
		},
		"/api/members/me/profile": {
			"get": {
				"operationId": "GetMyProfile",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IMember"
								}
							}
						}
					}
				},
				"tags": [
					"Member"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}